Fix: #106 Move some function from user to auth services
