+++ src/Services/auth.js
@@ -1,5 +1,15 @@
-import { LOCAL_STORAGE } from 'utils/constants';
+import { LOCAL_STORAGE, ROLES } from 'utils/constants';
+import { getCurrentUser } from './user';
 
 export const logout = () => {
   LOCAL_STORAGE.remove('token');
+};
+
+export const checkIsLoggedIn = () => {
+  return Boolean(getCurrentUser());
+};
+
+export const checkIsAdmin = () => {
+  const currentUser = getCurrentUser();
+  return currentUser.role === ROLES.ADMIN;
 };
